refactor: move cache_stack file
