+++ packages/bentocache/factories/cache_factory.ts
@@ -6,7 +6,7 @@ import { MemoryTransport } from '@boringnode/bus/transports/memory'
 import { Cache } from '../src/cache/cache.js'
 import { RedisDriver } from '../src/drivers/redis.js'
 import { MemoryDriver } from '../src/drivers/memory.js'
-import { CacheStack } from '../src/cache/stack/cache_stack.js'
+import { CacheStack } from '../src/cache/cache_stack.js'
 import { BentoCacheOptions } from '../src/bento_cache_options.js'
 import type { CacheStackDrivers, MemoryConfig } from '../src/types/main.js'
 import type { RawBentoCacheOptions } from '../src/types/options/options.js'
