+++ packages/bentocache/src/cache/get_set/single_tier_handler.ts
@@ -2,10 +2,10 @@ import type { MutexInterface } from 'async-mutex'
 
 import { Locks } from '../locks.js'
 import { errors } from '../../errors.js'
+import type { CacheStack } from '../cache_stack.js'
 import { FactoryRunner } from '../factory_runner.js'
 import type { Factory } from '../../types/helpers.js'
 import type { CacheEvent } from '../../types/events.js'
-import type { CacheStack } from '../stack/cache_stack.js'
 import { cacheEvents } from '../../events/cache_events.js'
 import type { CacheEntry } from '../cache_entry/cache_entry.js'
 import type { CacheEntryOptions } from '../cache_entry/cache_entry_options.js'
